modified:   makefile 	modified:   src/communication/io_service_pool.cpp 	modified:   src/communication/io_service_pool.h 	new file:   src/controllers/device_cassandra_controller.cpp 	new file:   src/controllers/device_cassandra_controller.h 	modified:   src/controllers/home_agent_index_cassandra_controller.cpp 	modified:   src/database/cassandra_db.cpp 	new file:   src/datastructure/configuration.h 	modified:   src/global.h 	new file:   src/models/device.cpp 	new file:   src/models/device.h 	new file:   src/models/user.cpp 	modified:   src/models/user.h 	modified:   src/protocol/dns/datastructure/dns_message.h 	modified:   src/protocol/dns/datastructure/dns_resource_record.h 	modified:   src/protocol/dns/dns_message_handler.cpp 	modified:   src/protocol/dns/dns_message_handler.h 	new file:   src/pweb_home_agent.cpp 	modified:   src/server/home_agent_server.h 	modified:   src/server/udp_connection.cpp 	modified:   src/server/udp_connection.h 	renamed:    src/test/pweb_home_agent_test.cpp -> src/test/pweb_home_agent_test.cpp.bak

+++ src/protocol/dns/datastructure/dns_message.h
@@ -121,20 +121,20 @@ public:
 	{
 		printf("------------------------------------------\n");
 		header.print();
-		printf("nQuestions: %d\n", questions.size());
-		for(int i = 0; i < questions.size(); i++)
+		printf("nQuestions: %lu\n", questions.size());
+		for(int i = 0; i < (int)questions.size(); i++)
 			questions[i].print();
 
-		printf("nAnswers: %d\n", answers.size());
-		for(int i = 0; i < answers.size(); i++)
+		printf("nAnswers: %lu\n", answers.size());
+		for(int i = 0; i < (int)answers.size(); i++)
 			answers[i].print();
 
-		printf("nAuthority: %d\n", authority.size());
-		for(int i = 0; i < authority.size(); i++)
+		printf("nAuthority: %lu\n", authority.size());
+		for(int i = 0; i < (int)authority.size(); i++)
 			authority[i].print();
 
-		printf("nAdditional: %d\n", additional.size());
-		for(int i = 0; i < additional.size(); i++)
+		printf("nAdditional: %lu\n", additional.size());
+		for(int i = 0; i < (int)additional.size(); i++)
 			additional[i].print();
 		printf("------------------------------------------\n");
 	}
